+++ CassoCore/Pch.h
@@ -47,6 +47,7 @@
 #include <print>
 #include <memory>
 #include <optional>
+#include <set>
 #include <span>
 #include <sstream>
 #include <string>